Refactor config option and tests
