+++ instrumentation/spring-4.3.0/src/main/java/com/nr/agent/instrumentation/SpringControllerUtility.java
@@ -26,7 +26,8 @@ import java.util.logging.Level;
 public class SpringControllerUtility {
     private static final String CGLIB_CLASS_SUFFIX = "$$EnhancerBy";
 
-    public static final boolean ENHANCED_NAMING_ENABLED =
+    // Package visibility for unit tests
+    static boolean ENHANCED_NAMING_ENABLED =
             NewRelic.getAgent().getConfig().getValue("class_transformer.enhanced_spring_transaction_naming", false);
 
     /**
