+++ instrumentation/spring-webflux-controller-mappings-5.0.0/src/test/java/org/springframework/web/reactive/result/method/TestControllerClasses.java
@@ -4,7 +4,7 @@
  *  * SPDX-License-Identifier: Apache-2.0
  *
  */
-package com.nr.agent.instrumentation.web.reactive;
+package org.springframework.web.reactive.result.method;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
